backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.1
motor==3.3.1
redis==5.0.1
aiohttp==3.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
emails==0.6
aiosmtplib==2.0.2
pytest==7.4.3
pytest-asyncio==0.21.1
prometheus-client==0.19.0
APScheduler==3.10.4
python-dotenv==1.0.0
websockets==12.0
